feat: TOC를 잉크 미니맵 플로팅으로 통일
