+++ src/components/post/index.tsx
@@ -9,7 +9,7 @@ interface PostProps {
 
 const Post = ({ toc, MDX }: PostProps) => {
   return (
-    <section className="relative space-y-12">
+    <section>
       {toc.length > 0 && <TOC toc={toc} />}
       <article className="space-y-6 *:first:mt-0 *:last:mb-0">
         <MDX />
