+++ src/static/styles/globals.css
@@ -21,6 +21,7 @@
 
   --animate-flip: flip 1s ease-in-out infinite;
   --animate-seal-press: seal-press 0.32s cubic-bezier(0.34, 1.56, 0.64, 1);
+  --animate-panel-in: panel-in 0.2s cubic-bezier(0.34, 1.3, 0.64, 1);
 
   @keyframes flip {
     0% {
@@ -42,6 +43,17 @@
       transform: rotate(-3deg) scale(1);
     }
   }
+
+  @keyframes panel-in {
+    0% {
+      opacity: 0;
+      transform: translateY(8px) scale(0.95);
+    }
+    100% {
+      opacity: 1;
+      transform: none;
+    }
+  }
 }
 
 [data-theme~='dark'] {
@@ -53,6 +65,14 @@
   --color-background: #171512;
   --color-accent: #f08c4b;
   --color-code-inline: #f0a36b;
+}
+
+@utility scrollbar-none {
+  scrollbar-width: none;
+
+  &::-webkit-scrollbar {
+    display: none;
+  }
 }
 
 @layer base {
